Merge new CE remote_mirror migrations into EE, and disable ones that are already in EE

See merge request gitlab-org/gitlab-ee!5606

+++ ee/db/post_migrate/20170427111108_remove_sync_time_column_from_remote_mirrors.rb
@@ -7,11 +7,11 @@ class RemoveSyncTimeColumnFromRemoteMirrors < ActiveRecord::Migration
 
   def up
     remove_concurrent_index :remote_mirrors, [:sync_time] if index_exists? :remote_mirrors, [:sync_time]
-    remove_column :remote_mirrors, :sync_time, :integer
+    remove_column :remote_mirrors, :sync_time, :integer if column_exists? :remote_mirrors, :sync_time
   end
 
   def down
-    add_column :remote_mirrors, :sync_time, :integer
-    add_concurrent_index :remote_mirrors, [:sync_time]
+    add_column :remote_mirrors, :sync_time, :integer unless column_exists? :remote_mirrors, :sync_time
+    add_concurrent_index :remote_mirrors, [:sync_time] unless index_exists? :remote_mirrors, [:sync_time]
   end
 end
